added a lot

+++ resources/views/menu/addMenuItems.blade.php
@@ -6,7 +6,7 @@
 	<div class="row justify-content-center">
 		
 		@if(session()->has('success'))
-		<div class="alert alert-success alert-dismissable fade show" role="alert">
+		<div class="alert alert-success alert-dismissable" role="alert">
 			{{session()->get('success')}}
 			<button type="button" class="close" data-dismiss="alert" aria-label="Close">
 				<span aira-hidden="true">&times;</span>
@@ -15,7 +15,7 @@
 		@endif
 
 		@if(session()->has('delete'))
-		<div class="alert alert-danger alert-dismissable fade show" role="alert">
+		<div class="alert alert-danger alert-dismissable" role="alert">
 			{{session()->get('delete')}}
 			<button type="button" class="close" data-dismiss="alert" aria-label="Close">
 				<span aira-hidden="true">&times;</span>
@@ -24,16 +24,39 @@
 		@endif
 
 		@if(session()->has('update'))
-		<div class="alert alert-secondary alert-dismissable fade show" role="alert">
+		<div class="alert alert-success alert-dismissable" role="alert">
 			{{session()->get('update')}}
 			<button type="button" class="close" data-dismiss="alert" aria-label="Close">
 				<span aira-hidden="true">&times;</span>
 			</button>
 		</div>
 		@endif
-
-
-
+		{{-- delete modal --}}
+		<div id="DeleteModal" class="modal fade text-danger" role="dialog">
+			<div class="modal-dialog ">
+				<!-- Modal content-->
+				<form action="" id="deleteForm" method="post">
+					<div class="modal-content">
+{{-- 						<div class="bg-danger">
+							<button type="button" class="close float-right" data-dismiss="modal">&times;</button>
+							<h4 class="modal-title text-center">DELETE CONFIRMATION</h4>
+						</div>
+ --}}						<div class="modal-body">
+							{{ csrf_field() }}
+							{{ method_field('DELETE') }}
+							<p class="text-center">Are you sure you want to delete this item?</p>
+						</div>
+						<div class="modal-footer">
+							<center>
+								<button type="button" class="btn btn-success" data-dismiss="modal">Cancel</button>
+								<button type="submit" name="" class="btn btn-danger" data-dismiss="modal" onclick="formSubmit()">Yes, Delete</button>
+							</center>
+						</div>
+					</div>
+				</form>
+			</div>
+		</div>
+		{{-- end of delete modal --}}
 	</div>
 	<div class="row">
 		<div class="col-md-3">
@@ -85,29 +108,52 @@
 								@csrf
 								<button>Update</button>
 							</form>
+							<a href="javascript:;" data-toggle="modal" onclick="deleteData({{$item->id}})" 
+								data-target="#DeleteModal" class="text-danger">Delete</a>
+								{{-- <a href="#" data-href="/deleteitem/{{$item->id}}" data-toggle="modal" data-target="#confirm-delete" class="text-danger">Delete</a> --}}
+							</td>
+						</tr>
+						@empty
+						<tr>
+							<td colspan="7" class="text-center">Nothing to show</td>
+						</tr>
 
-							<a href="/deleteitem/{{$item->id}}" class="text-danger">Delete</a>
-						</td>
-					</tr>
-					@empty
-					<tr>
-						<td colspan="7" class="text-center">Nothing to show</td>
-					</tr>
 
-					@endforelse
 
-				</tbody>
-				
-			</table>
+						@endforelse
+
+					</tbody>
+
+				</table>
+			</div>
 		</div>
 	</div>
-</div>
 
+
+	<script
+	src="https://code.jquery.com/jquery-3.4.1.min.js"
+	integrity="sha256-CSXorXvZcTkaix6Yvo6HppcZGetbYMGWSFlBw8HfCJo="
+	crossorigin="anonymous">
+</script>
 
 
 <script type="text/javascript">
-	$('#myModal').on('shown.bs.modal', function () {
-		$('#myInput').trigger('focus')
-	})
+	$(document).ready(function(){
+		$('.alert').delay(5000).fadeOut(300);
+	});
+
+     function deleteData(id)
+     {
+         var id = id;
+         var url = "/deleteitem/"+":id";
+         url = url.replace(':id', id);
+         $("#deleteForm").attr('action', url);
+     }
+
+     function formSubmit()
+     {
+         $("#deleteForm").submit();
+     }
+
 </script>
 @endsection
